Fixed SNX extender page parsing with BeautifulSoup 4.9.x
BeautifulSoup 4.9.x represents the contents of the <script> tag
with a bs4.element.Script object, which inherits from the
bs4.element.NavigableString. However, these objects are not
captured by default types=(NavigableString, CData) filter in
.text property (and underlying get_text() method), because
inheritance is not verified, only the type itself.

Therefore, trying to access script element's .text results in
an empty string, and no extender data is collected. To fix this,
we now use .get_text() with explicitly specified types, which
include bs4.element.Script if available.